newstory: small cleanup

+++ plugins/NewStory/src/utils.cpp
@@ -22,11 +22,10 @@ void CopyText(HWND hwnd, const wchar_t *text)
 {
 	OpenClipboard(hwnd);
 	EmptyClipboard();
-	HGLOBAL hMem = GlobalAlloc(GMEM_MOVEABLE, sizeof(wchar_t) * (int)(mir_wstrlen(text) + 1));
+	HGLOBAL hMem = GlobalAlloc(GMEM_MOVEABLE, sizeof(wchar_t) * (mir_wstrlen(text) + 1));
 	wchar_t *s = (wchar_t *)GlobalLock(hMem);
 	mir_wstrcpy(s, text);
 	GlobalUnlock(hMem);
 	SetClipboardData(CF_UNICODETEXT, hMem);
 	CloseClipboard();
-	//	GlobalFree(hMem);
 }
